Adding backend validations and removing inline form for parent AwardTemplate

+++ nominate_app/views/award_template_views.py
@@ -18,18 +18,20 @@ def new_award_template(request, award_id):
     is_active_val = new_form.get('is_active',False)
     is_active = True if is_active_val ==  'on' else False
     award_id = award.id
-    award_template = AwardTemplate.objects.create(template_name= template_name, is_active=is_active, award_id=award_id)
-    request.POST = new_form
-    formset = AwardTemplateFormset(request.POST)
-    if formset.is_valid():
-      instances = formset.save(commit=False)
-      for obj in formset.deleted_objects:
-        obj.delete()
-      for instance in instances:
-        instance.award_template_id=award_template.id
-        instance.save()
-      return redirect('nominate_app:award_template_index')
-
+    if (template_name != ''):
+      award_template =AwardTemplate.objects.create(template_name= template_name, is_active=is_active, award_id=award_id)
+      request.POST = new_form
+      formset = AwardTemplateFormset(request.POST)
+      if formset.is_valid():
+        instances = formset.save(commit=False)
+        for obj in formset.deleted_objects:
+          obj.delete()
+        for instance in instances:
+          instance.award_template_id=award_template.id
+          instance.save()
+        return redirect('nominate_app:award_template_index')
+    else:
+      return render(request, 'nominate_app/new_award_template.html', {'formset':formset, 'template_name': award.name, 'form_errors': 'Template name is required'})
   else:
     formset = AwardTemplateFormset(queryset=Questions.objects.none())
